requirements.txt
dj_database_url==0.5.0
requests==2.11.1
Django==2.2
numpy==1.16.4
matplotlib==3.1.0 
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2019.1
whitenoise==4.1.2
djangorestframework==3.9.4
django-cors-headers==3.0.2
ChatterBot==1.0.5
chatterbot-corpus==1.2.0